requirements.txt
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
torch>=2.4.0
torchvision>=0.19.0
sentencepiece>=0.2.0
protobuf>=5.27.0
runpod>=1.6.0
pillow>=10.4.0